+++ components/faphub/dao/network/src/main/java/com/flipperdevices/faphub/dao/network/retrofit/model/detailed/RetrofitSourceCodeLinks.kt
@@ -1,0 +1,10 @@
+package com.flipperdevices.faphub.dao.network.retrofit.model.detailed
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RetrofitSourceCodeLinks(
+    @SerialName("type") val type: String,
+    @SerialName("uri") val url: String
+)