+++ components/faphub/installedtab/impl/src/main/java/com/flipperdevices/faphub/installedtab/impl/viewmodel/InstalledFapsViewModel.kt
@@ -2,6 +2,8 @@ package com.flipperdevices.faphub.installedtab.impl.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.flipperdevices.core.log.LogTagProvider
+import com.flipperdevices.core.log.error
 import com.flipperdevices.faphub.dao.api.FapNetworkApi
 import com.flipperdevices.faphub.dao.api.model.SortType
 import com.flipperdevices.faphub.installedtab.impl.model.FapInstalledScreenState
@@ -13,22 +15,30 @@ import tangle.viewmodel.VMInject
 
 class InstalledFapsViewModel @VMInject constructor(
     private val fapNetworkApi: FapNetworkApi
-) : ViewModel() {
+) : ViewModel(), LogTagProvider {
+    override val TAG = "InstalledFapsViewModel"
     private val fapInstalledScreenStateFlow = MutableStateFlow<FapInstalledScreenState>(
         FapInstalledScreenState.Loading
     )
 
     init {
-        viewModelScope.launch {
-            val faps = fapNetworkApi.getAllItem(sortType = SortType.UPDATED, offset = 0, limit = 4)
-            fapInstalledScreenStateFlow.emit(
-                FapInstalledScreenState.Loaded(
-                    faps.toImmutableList()
-                )
-            )
-        }
+        onRefresh()
     }
 
     fun getFapInstalledScreenState(): StateFlow<FapInstalledScreenState> =
         fapInstalledScreenStateFlow
+
+    fun onRefresh() = viewModelScope.launch {
+        fapNetworkApi.getAllItem(sortType = SortType.UPDATE_AT_DESC, offset = 0, limit = 4)
+            .onSuccess { faps ->
+                fapInstalledScreenStateFlow.emit(
+                    FapInstalledScreenState.Loaded(
+                        faps.toImmutableList()
+                    )
+                )
+            }.onFailure {
+                error(it) { "Failed get installed fap" }
+                fapInstalledScreenStateFlow.emit(FapInstalledScreenState.Error(it))
+            }
+    }
 }