+++ components/faphub/maincard/impl/src/main/java/com/flipperdevices/faphub/maincard/impl/viewmodel/FapMainCardViewModel.kt
@@ -20,9 +20,7 @@ class FapMainCardViewModel @VMInject constructor(
 
     init {
         viewModelScope.launch {
-            runCatching {
-                fapNetworkApi.getFeaturedItem()
-            }.onSuccess {
+            fapNetworkApi.getFeaturedItem().onSuccess {
                 fapMainCardStateFlow.emit(FapMainCardState.Loaded(it))
             }.onFailure {
                 error(it) { "Failed load suggested item" }